Basic exception/gcsafe annotations
Both forward declarations and methods are assumed not gcsafe and raising
`Exception` unless otherwise specified.

This PR is a first pass over the API to ensure that it does not leak
unwanted `raises`/not-`gcsafe` effects into the application using
`nimqml` from the bulk of its API.

The changes do not touch the "public" QAIM methods but aims to fix at
least the internals. Of course, raising from within a QAIM callback
would crash the application but applying strict `raises: []` would be a
(significantly) breaking change since user code would now have to be
raises-correct.

* `c` functions do not raise exception (`importc` implies `gcsafe` but
not `raises:[]`.}
* `metaObject` instances are not suitable to expose as a global since
it's a ref type (accessing it from different threads is not gcsafe) -
this PR switches to a lazy threadvar instead

Free bonus: leak fix in QModelIndex